refactor(markdown): split example into types/inline/blocks/index
Amp-Thread-ID: https://ampcode.com/threads/T-019e6f0e-c7e6-704d-a34b-2b2db6b01e51
