+++ tests/examples/markdown/blocks.ts
@@ -1,0 +1,67 @@
+import {
+  seqC,
+  capture,
+  optional,
+  many1Till,
+  or,
+  manyTillStr,
+  count,
+  iManyTillStr,
+  many1,
+} from "@/lib/combinators";
+import { str, spaces, word, char, eof, set } from "@/lib/parsers";
+import { Parser, ParserResult, success } from "@/lib/types";
+import {
+  Heading,
+  CodeBlock,
+  BlockQuote,
+  Image,
+  Paragraph,
+} from "./types";
+import { inlineMarkdownParser } from "./inline";
+
+export const headingParser: Parser<Heading> = seqC(
+  set("type", "heading"),
+  capture(count(char("#")), "level"),
+  spaces,
+  capture(many1Till(or(char("\n"), eof)), "content")
+);
+
+export const codeBlockParser: Parser<CodeBlock> = seqC(
+  set("type", "code-block"),
+  str("```"),
+  capture(optional(word), "language"),
+  optional(spaces),
+  capture(manyTillStr("```"), "content"),
+  str("```")
+);
+
+export const blockQuoteParser: Parser<BlockQuote> = seqC(
+  set("type", "block-quote"),
+  str(">"),
+  spaces,
+  capture(manyTillStr("\n"), "content")
+);
+
+export const imageParser: Parser<Image> = seqC(
+  set("type", "image"),
+  str("!["),
+  capture(iManyTillStr("]("), "alt"),
+  str("]("),
+  capture(iManyTillStr(")"), "url"),
+  str(")")
+);
+
+export function paragraphParser(input: string): ParserResult<Paragraph> {
+  const inline = many1(inlineMarkdownParser)(input);
+  if (!inline.success) {
+    return inline;
+  }
+  return success(
+    {
+      type: "paragraph",
+      content: inline.result,
+    },
+    inline.rest
+  );
+}
