+++ tests/examples/markdown/index.ts
@@ -1,0 +1,32 @@
+export * from "./types";
+export * from "./inline";
+export * from "./blocks";
+export * from "./references";
+
+import { seq, sepBy, or, optional } from "@/lib/combinators";
+import { spaces } from "@/lib/parsers";
+import {
+  headingParser,
+  codeBlockParser,
+  blockQuoteParser,
+  paragraphParser,
+  imageParser,
+} from "./blocks";
+
+export const markdownParser = seq(
+  [
+    optional(spaces),
+    sepBy(
+      spaces,
+      or(
+        headingParser,
+        codeBlockParser,
+        blockQuoteParser,
+        paragraphParser,
+        imageParser
+      )
+    ),
+    optional(spaces),
+  ],
+  (r) => r[1]
+);
