+++ tests/examples/markdown/references.ts
@@ -1,0 +1,4 @@
+/* Reference link definitions and resolution pass.
+ * Will be populated in later phases of the implementation plan.
+ */
+export {};
